+++ server/impl/src/main/java/com/walmartlabs/concord/server/events/ExternalEventResource.java
@@ -27,7 +27,6 @@ import com.walmartlabs.concord.server.org.project.ProjectDao;
 import com.walmartlabs.concord.server.org.project.RepositoryDao;
 import com.walmartlabs.concord.server.org.triggers.TriggersDao;
 import com.walmartlabs.concord.server.process.ProcessManager;
-import com.walmartlabs.concord.server.security.ldap.LdapManager;
 import com.walmartlabs.concord.server.user.UserManager;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -65,10 +64,9 @@ public class ExternalEventResource extends AbstractEventResource implements Reso
                                  ProjectDao projectDao,
                                  RepositoryDao repositoryDao,
                                  TriggersConfiguration triggersCfg,
-                                 UserManager userManager,
-                                 LdapManager ldapManager) {
+                                 UserManager userManager) {
 
-        super(cfg, processManager, triggersDao, projectDao, repositoryDao, triggersCfg, userManager, ldapManager);
+        super(cfg, processManager, triggersDao, projectDao, repositoryDao, triggersCfg, userManager);
     }
 
     @POST