+++ server/impl/src/main/java/com/walmartlabs/concord/server/events/OneOpsEventResource.java
@@ -29,7 +29,6 @@ import com.walmartlabs.concord.server.org.project.ProjectDao;
 import com.walmartlabs.concord.server.org.project.RepositoryDao;
 import com.walmartlabs.concord.server.org.triggers.TriggersDao;
 import com.walmartlabs.concord.server.process.ProcessManager;
-import com.walmartlabs.concord.server.security.ldap.LdapManager;
 import com.walmartlabs.concord.server.user.UserManager;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.Authorization;
@@ -82,10 +81,9 @@ public class OneOpsEventResource extends AbstractEventResource implements Resour
                                ProjectDao projectDao,
                                RepositoryDao repositoryDao,
                                TriggersConfiguration triggersCfg,
-                               UserManager userManager,
-                               LdapManager ldapManager) {
+                               UserManager userManager) {
 
-        super(cfg, processManager, triggersDao, projectDao, repositoryDao, triggersCfg, userManager, ldapManager);
+        super(cfg, processManager, triggersDao, projectDao, repositoryDao, triggersCfg, userManager);
         this.objectMapper = new ObjectMapper();
     }