+++ server/impl/src/main/java/com/walmartlabs/concord/server/org/team/TeamDao.java
@@ -176,16 +176,17 @@ public class TeamDao extends AbstractDao {
     }
 
     public List<TeamUserEntry> listUsers(DSLContext tx, UUID teamId) {
-        return tx.select(USERS.USER_ID, USERS.USERNAME, USERS.USER_TYPE, USER_TEAMS.TEAM_ROLE)
+        return tx.select(USERS.USER_ID, USERS.USERNAME, USERS.DISPLAY_NAME, USERS.USER_TYPE, USER_TEAMS.TEAM_ROLE)
                 .from(USER_TEAMS)
                 .innerJoin(USERS).on(USERS.USER_ID.eq(USER_TEAMS.USER_ID))
                 .where(USER_TEAMS.TEAM_ID.eq(teamId))
                 .orderBy(USERS.USERNAME)
-                .fetch((Record4<UUID, String, String, String> r) ->
+                .fetch((Record5<UUID, String, String, String, String> r) ->
                         new TeamUserEntry(r.value1(),
                                 r.value2(),
-                                UserType.valueOf(r.value3()),
-                                TeamRole.valueOf(r.value4())));
+                                r.value3(),
+                                UserType.valueOf(r.value4()),
+                                TeamRole.valueOf(r.value5())));
     }
 
     public List<TeamLdapGroupEntry> listLdapGroups(UUID teamId) {