+++ server/impl/src/main/java/com/walmartlabs/concord/server/security/ldap/LdapGroupsDao.java
@@ -1,0 +1,68 @@
+package com.walmartlabs.concord.server.security.ldap;
+
+/*-
+ * *****
+ * Concord
+ * -----
+ * Copyright (C) 2017 - 2019 Walmart Inc.
+ * -----
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =====
+ */
+
+import com.walmartlabs.concord.db.AbstractDao;
+import org.jooq.BatchBindStep;
+import org.jooq.Configuration;
+import org.jooq.DSLContext;
+
+import javax.inject.Inject;
+import javax.inject.Named;
+import javax.inject.Singleton;
+import java.util.Set;
+import java.util.UUID;
+
+import static com.walmartlabs.concord.server.jooq.Tables.USER_LDAP_GROUPS;
+import static org.jooq.impl.DSL.value;
+
+@Singleton
+@Named
+public class LdapGroupsDao extends AbstractDao {
+
+    @Inject
+    public LdapGroupsDao(@Named("app") Configuration cfg) {
+        super(cfg);
+    }
+
+    public void update(UUID userId, Set<String> groups) {
+        tx(tx -> update(tx, userId, groups));
+    }
+
+    private void update(DSLContext tx, UUID userId, Set<String> groups) {
+        tx.deleteFrom(USER_LDAP_GROUPS).where(USER_LDAP_GROUPS.USER_ID.eq(userId))
+                .execute();
+
+        if (groups.isEmpty()) {
+            return;
+        }
+
+        BatchBindStep q = tx.batch(tx.insertInto(USER_LDAP_GROUPS, USER_LDAP_GROUPS.USER_ID, USER_LDAP_GROUPS.LDAP_GROUP)
+                .values((UUID) null, null));
+
+        for (String g : groups) {
+            q.bind(value(userId), value(g));
+        }
+
+        q.execute();
+    }
+
+}